pulp_func (exe_start): Fix logger enable bug and make code more robust
Previous to this patch, the value of `pulp->rab_ax_log_en` overwrote the
cluster select in the GPIO pin.  This was wrong as it did not enable the
logger, could disable the cluster select, and could even enable the wrong
clusters if `rab_ax_log_en` was uninitialized and had one of particular
values.

This patch has three main effects:
1) values of `rab_ax_log_en` other than `1` (that may occur with
   uninitialized variables) do not have any effect at all,
2) the RAB AX Loggers are enabled if `rab_ax_log_en` is set to 1, and
3) `pulp->cluster_sel` only affects the GPIO bits that are meant to
   activate clusters,