pulp_rab (soc_mh_ena): Remove writing of PGD PA to SoC
As mentioned in the comment also deleted with this commit, communicating
the physical address of the PGD to the SoC is unnecessary and thus only
clutters the mailbox.  This is because the SoC never uses the physical
address of the first-level page table (i.e., the PGD), but only accesses
the PGD through the RAB mapping setup by the driver in this function.

Instead, a single core should be programmed to handle RAB misses through
the program that is offloaded to the SoC.