pulp_rab (soc_mh_ena): Map the first level of the Page Table in Slice 1
Previous to this commit, the first level of the Page Table was always
set up at Slice 4.  The reason for this lies in legacy code and use
cases that are no longer applicable.  With this commit, the first level
of the Page Table is mapped in Slice 1, directly after the contiguous L3
memory mapping in Slice 0.