pulp_rab (soc_mh_ena): Fix assignment of 'page_ptr_idx'
Previous to this commit, the 'page_ptr_idx' field of a slice request was
manually set to a value that would (hopefully) not collide with other
slices that were already set up.  As of this patch, the 'page_ptr_idx'
field of a slice request is obtained through the 'page_ptrs_get_field()'
function, which is how the API was designed to work.