pulp_rab (soc_mh_ena): Fix assignment of 'rab_slice'
Previous to this commit, the 'rab_slice' field of a slice request was
manually set to a value that would (hopefully) not collide with other
slices that were already set up.  As of this patch, the 'rab_slice'
field of a slice request is obtained through the 'pulp_rab_slice_get()'
function, which is how the API was designed to work.