pulp_smmu.*: improve fault handling
This commit disables context fault interrupts before scheduling the bottom
half handler. The bottom half handler then triggers the retry of the
faulting transaction and re-enables context fault interrupts.

Compared to the previous proof-of-concept implementation, this is much
more reliable. The spin-lock is just needed to synchronize on the iova.