move macros and constants from `pulp_host.h` to driver and runtime files
For the driver, most of the macros and constants, e.g., such related to the
memory map are now found inside `pulp_module.h`. If appropriate, they are
now found inside submodule-specific headers such as `pulp_rab.h`.

For the user-space runtime library, the macros and constants are now
defined inside `pulp_func.h`.

This strict separation is required for the open-source release, as we
cannot share any driver code (GPLv2) and user-space or SDK code (not GPLv2,
Apache 2). In the future, we need to manually synchronize kernel-level
headers and user-space headers on interface changes (IOTCL, mailbox
signaling).