sourceme.sh: Remove definition of `PULP_CURRENT_CONFIG`
The `PULP_CURRENT_CONFIG` environment variable is set by the PULP SDK to
match the configuration the SDK is currently built for.  Setting it in
this repository will cause problems.  There are two cases that could
happen:

1.  This `sourceme.sh` is sourced *before* that of the SDK, causing the
    value of the SDK to override the one in here.  In that case, the
    value in here has no effect, and changing it does not produce the
    expected result.

2.  This `sourceme.sh` is sourced *after* that of the SDK, and the value
    in here overrides that of the SDK.  In this case, the `pulp-tools`
    (e.g., `plpbuild`), which rely on a value and format defined by the
    SDK, will no longer work properly.

To avoid this, this commit removes the definition of
`PULP_CURRENT_CONFIG` from this repository.