pulp_rab.c: switch to generic access function for `pmd_t`
`pmd_t` can be a struct or an unsigned 64bit type depending on the
kernel config. To be safe, it should be accessed through a generic access
function, see `arch/arm64/include/asm/pgtable-types.h` in the kernel dir.

This commit is equal to `63c72874dad5e5250a7669d060541a11009dbdc2` on the
old internal repo.